Refactoring: Extract HttpRequest extensions
