+++ BTCPayServer/Controllers/UIAppsController.PointOfSale.cs
@@ -6,6 +6,7 @@ using System.Text.Encodings.Web;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
+using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Models.AppViewModels;
 using BTCPayServer.Services.Apps;
 using Microsoft.AspNetCore.Mvc;
