+++ BTCPayServer/Views/Shared/LNURL/CrowdfundLNURL.cshtml
@@ -1,5 +1,6 @@
 ﻿@using BTCPayServer.Services.Stores
 @using BTCPayServer.Payments.Lightning
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Models.AppViewModels.ViewCrowdfundViewModel
 @inject BTCPayNetworkProvider BTCPayNetworkProvider
 @inject StoreRepository StoreRepository
