+++ BTCPayServer/Views/Shared/_Footer.cshtml
@@ -1,4 +1,5 @@
-﻿@inject BTCPayServer.Services.BTCPayServerEnvironment _env
+﻿@using BTCPayServer.Abstractions.Extensions
+@inject BTCPayServer.Services.BTCPayServerEnvironment _env
 
 <footer class="btcpay-footer">
     <div class="container">
