+++ BTCPayServer/Views/Shared/_LayoutPos.cshtml
@@ -3,6 +3,7 @@
 
 @using BTCPayServer.Services.Apps
 @using BTCPayServer.Abstractions.Contracts
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Models.AppViewModels.ViewPointOfSaleViewModel
 @{
     ViewData["Title"] = Model.Title;
