+++ BTCPayServer/Views/UIPullPayment/ViewPullPayment.cshtml
@@ -1,5 +1,6 @@
 ﻿@using NUglify.Helpers
 @using BTCPayServer.Abstractions.Contracts
+@using BTCPayServer.Abstractions.Extensions
 @model BTCPayServer.Models.ViewPullPaymentModel
 
 @addTagHelper *, BundlerMinifier.TagHelpers
